Revert "`Usability`: Change alert style (#315)"
This reverts commit 78f7e82c0295e54b316a241b395180a393504909.

+++ src/main/webapp/app/application/application-overview-for-applicant/application-overview-for-applicant.component.ts
@@ -3,8 +3,6 @@ import { Router } from '@angular/router';
 import { ApplicationOverviewDTO, ApplicationResourceService } from 'app/generated';
 import { ButtonComponent } from 'app/shared/components/atoms/button/button.component';
 import { DynamicTableColumn, DynamicTableComponent } from 'app/shared/components/organisms/dynamic-table/dynamic-table.component';
-import { ToastComponent } from 'app/shared/toast/toast.component';
-import { ToastService } from 'app/service/toast-service';
 import { TableLazyLoadEvent } from 'primeng/table';
 import { firstValueFrom } from 'rxjs';
 import { BadgeModule } from 'primeng/badge';
@@ -15,7 +13,7 @@ import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'jhi-application-overview-for-applicant',
-  imports: [DynamicTableComponent, ButtonComponent, BadgeModule, SharedModule, TranslateModule, ToastComponent],
+  imports: [DynamicTableComponent, ButtonComponent, BadgeModule, SharedModule, TranslateModule],
   templateUrl: './application-overview-for-applicant.component.html',
   styleUrl: './application-overview-for-applicant.component.scss',
 })
@@ -78,7 +76,6 @@ export default class ApplicationOverviewForApplicantComponent {
   });
 
   private readonly router = inject(Router);
-  private toastService = inject(ToastService);
 
   private readonly applicationService = inject(ApplicationResourceService);
   private readonly accountService = inject(AccountService);
@@ -132,20 +129,21 @@ export default class ApplicationOverviewForApplicantComponent {
   }
 
   onDeleteApplication(applicationId: string): void {
-    // TODO nicer looking confirm, add dialog
-    // if (confirmDelete) {
-    this.applicationService.deleteApplication(applicationId).subscribe({
-      next: () => {
-        this.toastService.showSuccess({ detail: 'Application successfully deleted' });
-        const event = this.lastLazyLoadEvent();
-        if (event) this.loadPage(event);
-      },
-      error: err => {
-        this.toastService.showError({ detail: 'Error withdrawing the application' });
-        console.error('Delete failed', err);
-      },
-    });
-    // }
+    // TODO nicer looking confirm
+    const confirmDelete = confirm('Do you really want to delete this application?');
+    if (confirmDelete) {
+      this.applicationService.deleteApplication(applicationId).subscribe({
+        next: () => {
+          alert('Application successfully deleted');
+          const event = this.lastLazyLoadEvent();
+          if (event) this.loadPage(event);
+        },
+        error(err) {
+          alert('Error withdrawing the application');
+          console.error('Delete failed', err);
+        },
+      });
+    }
   }
 
   onWithdrawApplication(applicationId: string): void {
@@ -154,12 +152,12 @@ export default class ApplicationOverviewForApplicantComponent {
     if (confirmWithdraw) {
       this.applicationService.withdrawApplication(applicationId).subscribe({
         next: () => {
-          this.toastService.showSuccess({ detail: 'Application successfully withdrawn' });
+          alert('Application successfully withdrawn');
           const event = this.lastLazyLoadEvent();
           if (event) this.loadPage(event);
         },
-        error: err => {
-          this.toastService.showError({ detail: 'Error withdrawing the application' });
+        error(err) {
+          alert('Error withdrawing the application');
           console.error('Withdraw failed', err);
         },
       });
